basic-agent/requirements.txt
langchain
langgraph
langchain-google-genai
python-dotenv